Re-added all Portfolio files.
I was able to solve the issue with the toggle light/dark mode button.  My updates were not recorded within the folder in my computer, so I had to delete all of the files and make new ones. This means that all of my previous pushes are gone as well.